+++ apps/dotcom/sync-worker/src/routes/tla/forwardRoomRequest.ts
@@ -1,8 +1,8 @@
 import { ROOM_PREFIX } from '@tldraw/dotcom-shared'
 import { notFound } from '@tldraw/worker-shared'
 import { IRequest } from 'itty-router'
-import { Environment } from '../types'
-import { isRoomIdTooLong, roomIdIsTooLong } from '../utils/roomIdIsTooLong'
+import { Environment } from '../../types'
+import { isRoomIdTooLong, roomIdIsTooLong } from '../../utils/roomIdIsTooLong'
 
 // Forwards a room request to the durable object associated with that room
 export async function forwardRoomRequest(request: IRequest, env: Environment): Promise<Response> {
