+++ apps/dotcom/sync-worker/src/worker.ts
@@ -10,20 +10,20 @@ import { createRouter, handleApiRequest, notFound } from '@tldraw/worker-shared'
 import { WorkerEntrypoint } from 'cloudflare:workers'
 import { cors } from 'itty-router'
 import { APP_ID } from './TLAppDurableObject'
-import { createFiles } from './routes/createFiles'
 import { createRoom } from './routes/createRoom'
 import { createRoomSnapshot } from './routes/createRoomSnapshot'
-import { duplicateRoom } from './routes/duplicateFile'
 import { extractBookmarkMetadata } from './routes/extractBookmarkMetadata'
-import { forwardRoomRequest } from './routes/forwardRoomRequest'
-import { getPublishedFile } from './routes/getPublishedFile'
 import { getReadonlySlug } from './routes/getReadonlySlug'
 import { getRoomHistory } from './routes/getRoomHistory'
 import { getRoomHistorySnapshot } from './routes/getRoomHistorySnapshot'
 import { getRoomSnapshot } from './routes/getRoomSnapshot'
 import { joinExistingRoom } from './routes/joinExistingRoom'
-import { publishFile } from './routes/publishFile'
-import { unpublishFile } from './routes/unpublishFile'
+import { createFiles } from './routes/tla/createFiles'
+import { duplicateRoom } from './routes/tla/duplicateFile'
+import { forwardRoomRequest } from './routes/tla/forwardRoomRequest'
+import { getPublishedFile } from './routes/tla/getPublishedFile'
+import { publishFile } from './routes/tla/publishFile'
+import { unpublishFile } from './routes/tla/unpublishFile'
 import { Environment } from './types'
 import { getAuth } from './utils/tla/getAuth'
 export { TLAppDurableObject } from './TLAppDurableObject'
